armada-370-rd: use eth0 interface
The Armada 370 RD is a bit special:

 - It has a WAN interface, known as eth0 under Linux
 - It has a switch, connected to eth1 under Linux

Old Linux versions (such as Linux 3.16) did not support the switch, so
only eth0 is available in Linux.

However, U-Boot only supports network through the switch. So the
Armada 370 RD is connected in the board farm with two Ethernet cables:
one to the WAN interface, one to the switch. The port on the switch is
used by U-Boot, the WAN port is used by Linux.

This commit ensures that Linux uses the WAN interface.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>